commit to fix the way in which soil water is calculated and updated each day.
In the crop module, this meant refactoring soil_water_uptake.py to water_uptake.py because I kept confusing myself.

In the soil module, this meant deleting the updateCurrentSoilWater method from soil.py because it was messy, inconsistent with the spreadsheet model, and calculations shouldn't be performed in the driver class for the module in the first place.

The calculations previously performed in soil.py were cleaned up and moved into a separate class called soil_water.py. I left a few notes regarding outstanding questions for the calculation of daily soil water, but Missy is reviewing those and will be getting back to me shortly. Calculations for daily soil water are now included in the pseudocode for soil hydrology (section 1.D). These changes are also currently under review, but will ultimately be located in their finalized version in "pseudocode_SC_soilhydrology.docx" A number of commenting changes were also made to soil.py throughout the process.

The change to percolation.py was a commenting thing.