The following is a commit to donovan_barnyard_testing that encompasses numerous bug fixes that were caught while comparing output to the spreadsheet model. Using a new data set and a more generalized object for time exposed a number of interesting/unique errors.
barnyard.json has been updated in this commit because of an error in the end date that failed to account for 2012 being a leap year. The planting date was also updated to 367– this is a bootstrap method for turning off the crop routine while still generating values for use in the soil module.

programmer comments were removed in crop.py

in evapotranspiration.py, I removed programmer comments that no longer applied and removed the rounding of LAI to 3 digits– this was done in order to match output to the spreadsheet model, but RuFaS is more accurate, so this is unnecessary. I clarified variable names for Esoil calculations and hopefully made the whole process easier to understand– this is accompanied by an update in the pseudocode that is currently under review (05/02/2019)

I went through nitrogen_cycling.py and flagged areas that would need to be changed when I start working on nitrogen_cycling, but those changes will be performed in a separate branch.

commenting changes to soil.py including a fancy shmancy new header

fixed a bug in soil_erosion where snow adjusted sediment yield was being calculated but never used and removed a programmer comment

fixed a bug in soil_temp where Tsurf was being calculated using itself, rather than the temperature of the first layer of the soil

fixed a bug in classes.py that extended the year by one day at the end of the simulation

**IMPORTANT** changed the order of operations for the routine calls in simulation_engine.py– donovan_SW_fix updated SW to be calculated using an attribute from the Crop module– this attribute has to be calculated before the soil routine can run so now the daily crop routine runs before the daily soil routine. No crop is run for barnyard testing, so I will be interested to see how this changes the behavior of the crop routine.