Fixed issues with MP calculations using both NRC and NASEM
Issue with NASEM: A conversion in the metabolizable protein calculation was incorrectly applied. Specifically, we should divide the milk_true_protein value by 100 instead of the net proteins because that value is a percentage. That issue resulted in underestimation of MP using NASEM in RuFaS.

Issue with NRC: The equation for MPm (MP for maintenance) is incomplete. It is missing the dry matter intake and metabolizable bacterial protein part.