Update cohort2_farm4_diet.json
Substituted Aminoplus with soybean meal.

Updated feed cost for a few feeds.

Updated storage options. Uncertain about 'bunk_type' because it is not in the metadata; also not sure about 'initial_dry_matter' because it is in % when reported by the farmers. Noticed that some feed storage defaults are incorrectly entered in the default_animal.py file.