Updated c2f4 input files
Updated the animal input file for farm 4 in cohort 2 to match with the newest input format.
Updated the names of the animal, diet, and manure input files to match with the other farms.
Updated the pathname to the configuration file.
The herd structure looks good on this branch with the current animal inputs.