Update default_feed.json
Created a new default diet for all animal categories based on the FarmES NorthEast regional diets with the following adjustments.
(1) Replaced the mineral mixes (301) with minerals 231 and 234.
(2) Replaced the byproduct mixes with almond hulls, soybean meals, and corn grains with a ratio of 35%, 31% and 34% based on their ratios in the close-up cow diet from Edward's NE Diet 15.
(3) We further slightly adjusted the ratios of the three byproducts to allow for higher CP content in the diets.