Update default_animal.json
Updated the number of stalls so the code runs with no errors.

(1) Calculated the expected number of animals in each pen based on the inputs
(2) Multiply by 1.5 to accommodate the fluctuation in the number of animals during simulation.
(3) Round the value up to the nearest number that ends with '5' or '0'.