Fix Ration Interval Purchase
This PR fixes the ration interval feed purchase logic in `FeedManager.manage_ration_interval_purchases()`.
The previous logic was to purchase all the requested feed from the animal module, despite the current feed invnetory.
This PR goes through all of the requested feeds and compare the current inventory to find the actual amount that we need to purchase.