Remove `dummy_path` Generated by Unit Tests
This PR removes the generation of a `dummy_path/` directory when running the unit tests.
Specifically, when testing `HerdFactory.initialize_herd()` with `save_animals` set to True, we would call `om.create_directory()` to create a dummy directory. This PR patches the function, so that we don't actually create a directory.